Reworked NPC arousal behaviour
Changed NPC behaviour during arousal to communicate with potential sexnodes in range.
The nodes receive data from the NPC that is cross-checked with actors already present in the node.
Based on the compatibility the node receives a higher or lower score upon which the best node is chosen.
New config parameters to enable behaviour logging, change behaviour weights and disallow or punish incest.