Advanced arousal behavior
Navigation to sexnodes now aborts on three criterias: When the NPC doesn't actually move for about one second, it aborts. If the NPC doesn't get closer to the target for 30 seconds, it aborts. In both cases, node is flagged as unreachable, which ignores the node for 60 seconds.
Transformation nodes (AB LMB attack, defeat, ...) are now taken into consideration too.
Fixes #60