Fixed position switching
Fixed that the id of the "current" position is still set even when the actual switch is aborted due to the node being in the "nullState".
This fixes the GFY station not working with NPCs and, in the future, NPCs to successfully use solo-actor positions.
Also cleaned up irrelevant code.