Fixed position inconsistency on forced actor uninit
When using `uninitActors()` to flush a sexnode, the positions were never reset, as this bypasses the normal "remove actor" pipeline. This results in any attempt to add another actor afterwards, when it was flushed in non-idle state, to break. This is because the position filtering on the first added actor is done in null state, not actually switching to idle, never updating the current position pointer, which then points to an nil entry in the position list. This is now fixed.
Also added more failsafety to the system and reduced old commented out code, as well as reduced log spam on unnecessary duplicate "uninitActors" calls.