Lustia Rework
Rework of the Lustia dungeon, attempting to integrate addon content more naturally.
This version tries to add a closed, run-down version of the arcade from the Omnibrowser addon, with a clearer hint at to where and how to get the content.
Currently doesn't work, hard-crashes randomly and doesn't properly generate otherwise.