Multi Seat Lounging Logic
Changed the NPC aroused behavior to use new lounging logic that accounts for the amount of possible seats on the target object. This is needed because you *have* to define which lounge anchor to target with `world.lounge()`.
Seat count information is read from the object's config first orientation, cause objects with both orientations should never have differing amounts of seats.